Added List Items and SQL Database
Added intent transfers to the app to open the different activities.

Added the list for the list on the main screen

Added Ticket, TicketAdapter, TicketContract, TicketDbHandler for the SQL side of the app. Added ListItem class for the items in the list.

Still need to work on the TicketDbHandler to finish all of the operations to the database.